adding the mode condition to the from_scratch one

+++ src/koopmans/workflows/_dft.py
@@ -72,7 +72,7 @@ class DFTPWWorkflow(DFTWorkflow):
         calc.parameters.restart_mode = 'from_scratch'
 
         # Remove old directories
-        if self.parameters.from_scratch:
+        if self.parameters.from_scratch and calc.mode == 'ase':
             utils.system_call(f'rm -r {calc.parameters.outdir} 2>/dev/null', False)
 
         # Run the calculator
